test functionality now included in field power demo
Extra module/source code file no longer needed
Control of test code now via environment variable
